Created area generator using --> d3.area()
To draw the area between the minimum and the maximum temperatures, we use x(), y0(), and y1(). The first one calculates the horizontal position of each data point, the second the vertical position of the data points on the lower boundary (here, the minimum temperature), and the third one the vertical position of the data points on the upper edge (here, the maximum temperature).