Finished Arc chart of percentage of rainy days in NYC
Arcs are created with SVG path elements, for which the d attribute is calculated with an arc generator:
– D3’s arc generator (d3.arc()) has an accessor function that defines the start- ing and ending angle of an arc (startAngle() and endAngle()), as well as its inner and outer radius (innerRadius() and outerRadius()).
– We can also use accessor functions to round the corners of an arc (corner- Radius()) or to add padding between arcs (padAngle()).
– The arc generator expects angles to be expressed in radians.
The center of mass of an arc can be calculated with the centroid() method. Chained to an arc generator, this accessor function returns an array containing the horizontal and vertical position of the center of mass.