Implemented a working Stack generator  not finished
Starting with declaring a stack generator with method `d3.stack()` and stored it in a constant named `stackGenerator`

Then we tell the generator which keys from the dataset contain the values we want to stack (what will become the series)
- We do this with the `keys()` accessor function, to which we pass an array of the category IDs -- in this case, the identifier of each music format.
- We create this array by mapping the IDs from the `formatsInfo` constant

Next we use the annotated dataset returned by the stack layout to draw the bars corresponding to each category.

We'll also need a scale for the vertical axis to translate each rectangle's lower and upper boundary into a vertical position.
- We want the height of the bars to be linearly proportional to the sales, so we use a linear scale

This scale will need to access the annotated data, we'll declare it inside function `drawStackedBars()`
- the scale domain goes from 0 to the maximum upper boundary available in the annotated data.
- we know that this maximum value must live inside the last series of the annotated data, which will be positioned at the top of the chart

Accessing this series with the length property `annotatedData[annotatedData.length - 1]`
- Then we use method ` d3.max()` to retrieve the maximum value under property `d[1]`, which corresponds to the upper boundary

The range of the vertical scale varies from `innerHeight` the bottom of the inner chart, to 0, the top of the inner chart -( remember that the SVG vertical axis is positive in the downward direction)-
- we chain the scale declaration with method `.nice()` which will ensure that the domain ends on "nice" round values rather than the actual maximum value in the annotated dataset

Then we loop through each series from the annotated dataset
- Use the data-binding pattern to append a rectangle for each year.
- Apply a different class name to each series, and use it as a selector to avoid removing the previous rectangles as the loop progresses.

Finally use the scales to position the rectangles and set their fill attributes.